requirements.txt
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
lxml==5.3.0
more-itertools==10.6.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
simple-salesforce==1.12.6
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
zeep==4.3.1
